requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
supabase==2.10.0